pyproject.toml
[tool.poetry]
name = "pyharp"
version = "0.1.0"
description = "Library for data acquisition and control of devices implementing the Harp protocol."
authors = ["filcarv <filipe@open-ephys.org>"]
license = "MIT"
readme = 'README.md'
python-versions = '>3.10'

[tool.poetry.dependencies]
python = "^3.4"
pyserial = "^3.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.782"
black = "^19.10b0"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pyharp = "pyharp.main:main"
